Refactored prepare.js. Trying to follow more of a template design pattern
